requirements.txt
assemblyai==0.17.0
python-docx==0.8.11
openai==1.3.5
requests==2.31.0
google-cloud-speech==2.21.0
pyaudio==0.2.13 